Finally, lets give user opportunity to change server message. We add another route and corresponding controller action and attach another button to allow user input new server message.
After user input message, we post in to the server.
If server reports 'ok' - we just update content of message. This is a
way how we can avoid reloading page.

Note that now we have two routes on same path. Pne work with method
GET, it's a route to 'get' message, and other - with POST method, for
updating message content.

After we updated server message, we can reload page to be sure that
server message is really updated on server side.